// src/interface/LivroDTO.d.ts
export interface LivroDTO {
    id_livro?: number;
    titulo: string;
    autor: string;
    editora: string;
    ano_publicacao: number;
    isbn: number;
    quant_total: number;
    quant_disponivel: number;
    valor_aquisicao: number;
    status_livro_emprestado: number;
    situacao?: boolean;
}
//# sourceMappingURL=LivroDTO.d.ts.map